refactor: rules file name to id [skip ci]
