+++ src/modules/rules/correct-pages-connector.ts
