+++ src/modules/rules/correct-publication-title.ts
