+++ src/modules/rules/correct-thesis-type.ts
