+++ src/modules/rules/correct-university.ts
