+++ src/modules/rules/no-item-duplication.ts
