+++ src/modules/rules/no-title-trailing-dot.ts
