+++ src/modules/rules/require-abbr.ts
